avance nuevos filtros

+++ View/UltimosBeneficios/listar.php
@@ -1,6 +1,7 @@
 <p>Cantidad Registros:<b><?=$nro_reg?></b></p>
 <table class="table table-bordered">
     <tr>
+        <th>Nemonico</th>
         <th>Empresa</th>
         <th>Derechos</th>
         <th>Fecha de Acuerdo</th>
@@ -13,6 +14,7 @@
     while ($ub = mysqli_fetch_array($res)) {
 
         $nombre  = $ub['nombre'];
+        $ub_nemonico  = $ub['ub_nemonico'];
         $ub_der_mon  = $ub['ub_der_mon'];
         $ub_der_imp  = $ub['ub_der_imp'];
         $ub_der_por  = $ub['ub_der_por'];
@@ -26,6 +28,7 @@
         $derecho = $ub['ub_der_comp'];
     ?>
     <tr>
+        <td><?=$ub_nemonico?></td>
         <td><?=$nombre?></td>
         <td><?=$derecho?></td>
         <td><?=$ub_fech_acu?></td>
@@ -34,7 +37,7 @@
         <td><?=$ub_fech_ent?></td>
         <td>
             <a href="../Controller/UltimosBeneficiosC.php?accion=edit&ub_cod=<?=$ub['ub_cod']?>" class="" role="button"><i class="fa fa-pencil-square-o fa-2x" aria-hidden="true"></i></a>&nbsp;
-		    <a href="../Controller/UltimosBeneficiosC.php?accion=delete&ub_cod=<?=$ub['ub_cod']?>" class="color-red" role="button"><i class="fa fa-trash-o fa-2x" aria-hidden="true"></i></a>
+		    <a href="../Controller/UltimosBeneficiosC.php?accion=delete&ub_cod=<?=$ub['ub_cod']?>" class="color-red" role="button" onclick="return confirm('Esta seguro de eliminar el registro seleccionado?')"><i class="fa fa-trash-o fa-2x" aria-hidden="true"></i></a>
         </td>
     </tr>
     <?php
